Added in a default configuration file for Xillinux.
This replaces the first patch in the Xillinux patch set for the linux
kernel. Rather than having a patch that generates a .config file for the
kernel, this uses the existing kernel build system, putting the
configuration settings instead in the default configuration file
directory. It can then be accessed with `make xillinux_zynq_defconfig`.