This commit enables the use of a usb ethernet gadget.
A USB ethernet gadget allows one to obtain an ethernet connection over a
USB cable with the host computer. This changes the USB OTG port on the
Zedboard (J10) to be in peripheral mode, which allows for gadgets to be
deployed.

You can enable the USB ethernet gadget by inserting the following
modules:
```
modprobe libcomposite
modprobe u_ether
modprobe usb_f_rndis
modprobe g_ether
```

This will then pop up a USB0 network interface on both the zedboard and
the host computer.