Added in cyclic scatter-gather support into the Xilinx DMA driver.
The cyclic mode of the AXI DMA IP is when the DMA repeatedly reads the
same scatter-gather list over and over again, ignoring the completed
bits (which would normally lead to stale descriptor errors). The scatter
list's last element is chaned to the first, so the DMA engine repeatedly
walks over the list, and when it reaches back to the end, it cycles back
to the first descriptor.

This is implemented by extending Xilinx's configuration structure with a
boolean indicating if the current user of the engine wants a cyclic
transfer. If so, the driver changes the programming of the AXI DMA
appropiately.